6Zp8T2LTKT-C1: Merge every small http response file into http response manager.
