+++ src/External/Presentation/GoalFinder.WebApi/Endpoints/Auth/ForgotPassword/HttpResponseMapper/ForgotPasswordHttpResponseManager.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System;
 using GoalFinder.Application.Features.Auth.ForgotPassword;
+using Microsoft.AspNetCore.Http;
 
-namespace GoalFinder.WebApi.Endpoints.Auth.ForgotPassword.HttpResponseMapper.Others;
+namespace GoalFinder.WebApi.Endpoints.Auth.ForgotPassword.HttpResponseMapper;
 
 /// <summary>
 ///     Mapper for forgot password
@@ -24,23 +25,44 @@ internal sealed class ForgotPasswordHttpResponseManager
 
         _dictionary.Add(
             key: ForgotPasswordReponseStatusCode.OPERATION_SUCCESS,
-            value: (_, response) => new OperationSuccessHttpResponse(response: response));
+            value: (_, response) => new()
+            {
+                HttpCode = StatusCodes.Status200OK,
+                AppCode = response.StatusCode.ToAppCode(),
+                Body = response.ResponseBody
+            });
 
         _dictionary.Add(
             key: ForgotPasswordReponseStatusCode.USER_WITH_EMAIL_IS_NOT_FOUND,
-            value: (_, response) => new UserWithEmailNotFoundHttpsResponse(response: response));
+            value: (_, response) => new()
+            {
+                HttpCode = StatusCodes.Status404NotFound,
+                AppCode = response.StatusCode.ToAppCode()
+            });
 
         _dictionary.Add(
             key: ForgotPasswordReponseStatusCode.INPUT_VALIDATION_FAIL,
-            value: (_, response) => new InputValidationFailHttpResponse(response: response));
+            value: (_, response) => new()
+            {
+                HttpCode = StatusCodes.Status400BadRequest,
+                AppCode = response.StatusCode.ToAppCode()
+            });
 
         _dictionary.Add(
             key: ForgotPasswordReponseStatusCode.USER_IS_TEMPORARILY_REMOVED,
-            value: (_, response) => new UserIsTemporarilyRemovedHttpResponse(response: response));
+            value: (_, response) => new()
+            {
+                HttpCode = StatusCodes.Status403Forbidden,
+                AppCode = response.StatusCode.ToAppCode()
+            });
 
         _dictionary.Add(
             key: ForgotPasswordReponseStatusCode.DATABASE_OPERATION_FAIL,
-            value: (_, response) => new DatabaseOperationFailHttpResponse(response: response));
+            value: (_, response) => new()
+            {
+                HttpCode = StatusCodes.Status500InternalServerError,
+                AppCode = response.StatusCode.ToAppCode(),
+            });
     }
     /// <summary>
     /// Resolve
