--- src/External/Presentation/GoalFinder.WebApi/Endpoints/Auth/ForgotPassword/HttpResponseMapper/UserWithEmailNotFoundHttpsResponse.cs
@@ -1,16 +1,0 @@
-﻿using GoalFinder.WebApi.Endpoints.Auth.ForgotPassword.HttpResponseMapper.Others;
-using Microsoft.AspNetCore.Http;
-using GoalFinder.Application.Features.Auth.ForgotPassword;
-
-namespace GoalFinder.WebApi.Endpoints.Auth.ForgotPassword.HttpResponseMapper;
-/// <summary>
-///     User with email not found
-/// </summary>
-internal sealed class UserWithEmailNotFoundHttpsResponse : ForgotPasswordHttpReponse
-{
-    internal UserWithEmailNotFoundHttpsResponse(ForgotPasswordResponse response)
-    {
-        HttpCode = StatusCodes.Status404NotFound;
-        AppCode = response.StatusCode.ToAppCode();
-    }
-}
