+++ src/External/Presentation/GoalFinder.WebApi/Endpoints/Auth/Login/HttpResponseMapper/LazyLoginHttResponseMapper.cs
@@ -1,4 +1,4 @@
-﻿namespace GoalFinder.WebApi.Endpoints.Auth.Login.HttpResponseMapper.Others;
+﻿namespace GoalFinder.WebApi.Endpoints.Auth.Login.HttpResponseMapper;
 
 /// <summary>
 ///     login extension methods.
