+++ src/External/Presentation/GoalFinder.WebApi/Endpoints/Auth/RegisterAsUser/HttpResponseMapper/RegisterAsUserHttpResponse.cs
@@ -1,19 +1,19 @@
 using System.Collections.Generic;
 using System;
-using GoalFinder.Application.Features.Auth.Login;
 using System.Text.Json.Serialization;
+using GoalFinder.Application.Features.Auth.Register;
 
-namespace GoalFinder.WebApi.Endpoints.Auth.Login.HttpResponseMapper.Others;
+namespace GoalFinder.WebApi.Endpoints.Auth.RegisterAsUser.HttpResponseMapper;
 
 /// <summary>
-///     Implementation for login http response.
+///     Implementation for register as user http response.
 /// </summary>
-public class LoginHttpResponse
+internal sealed class RegisterAsUserHttpResponse
 {
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public int HttpCode { get; set; }
 
-    public string AppCode { get; init; } = LoginResponseStatusCode.OPERATION_SUCCESS.ToAppCode();
+    public string AppCode { get; init; } = RegisterAsUserResponseStatusCode.OPERATION_SUCCESS.ToAppCode();
 
     public DateTime ResponseTime { get; init; } = TimeZoneInfo.ConvertTimeFromUtc(
         dateTime: DateTime.UtcNow,
