+++ src/External/Presentation/GoalFinder.WebApi/Endpoints/Auth/RegisterAsUser/Middlewares/Caching/RegisterAsUserCachingPostProcessor.cs
@@ -3,9 +3,9 @@ using GoalFinder.Application.Features.Auth.ForgotPassword;
 using GoalFinder.Application.Features.Auth.Login;
 using GoalFinder.Application.Features.Auth.Register;
 using GoalFinder.Application.Shared.Caching;
-using GoalFinder.WebApi.Endpoints.Auth.Login.HttpResponseMapper.Others;
+using GoalFinder.WebApi.Endpoints.Auth.Login.HttpResponseMapper;
 using GoalFinder.WebApi.Endpoints.Auth.RegisterAsUser.Common;
-using GoalFinder.WebApi.Endpoints.Auth.RegisterAsUser.HttpResponseMapper.Others;
+using GoalFinder.WebApi.Endpoints.Auth.RegisterAsUser.HttpResponseMapper;
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Threading;
